+++ src/components/auth/index.ts
@@ -1,2 +1,4 @@
 export { AuthProvider, useAuth } from "./AuthProvider";
 export { AuthGuard } from "./AuthGuard";
+export { useInviteAcceptance, type InvitationStatus } from "./useInviteAcceptance";
+export { RoleSelector } from "./RoleSelector";
